+++ database/migrations/2023_03_27_025727_create_calendar_event_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('calendar_events', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('brand_id')->nullable()->comment('브랜드 FK')->constrained('brands')->onDelete('SET NULL');
+            $table->unsignedSmallInteger('brand_id')->nullable()->comment('브랜드 FK')->constrained('brands')->onDelete('SET NULL');
             $table->integer('user_id')->comment('유저 ID');
             $table->tinyInteger('user_type')->default(0)->comment('유저 타입(0=일반유저, 1=가맹점, 2=영업자)');
             $table->string('title', 100)->comment('제목');
